requirements.txt
amqp==2.2.1
billiard==3.5.0.3
celery==4.1.0
certifi==2017.7.27.1
chardet==3.0.4
defusedxml==0.5.0
Django==1.11.3
django-allauth==0.32.0
django-braces==1.11.0
django-celery-beat==1.0.1
django-celery-results==1.0.1
django-crispy-forms==1.6.1
django-debug-toolbar==1.8
django-model-utils==3.0.0
django-paypal==0.3.6
djangorestframework==3.6.3
ebaysdk==2.1.4
gunicorn==19.7.1
idna==2.5
jsonfield==2.0.2
kombu==4.1.0
lxml==3.8.0
oauthlib==2.0.2
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.1
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2017.2
redis==2.10.5
requests==2.18.3
requests-oauthlib==0.8.0
six==1.10.0
sqlparse==0.2.3
stripe==1.62.0
Unipath==1.1
urllib3==1.22
vine==1.1.4
